# Improved prompts for better consistent results in workflow.py