## Context

<!--
In a few sentences or less, please explain the context behind this change to help answer why this change is needed.

If this is a bug fix, make sure to include "fixes #xxxx", or
"closes #xxxx".

Screenshots, logs, code or other visual aids are greatly appreciated.
 -->
I was not able to get the desired result consistently in the required format. The GGUF model sometimes attached the prompt and some annotations.  So I did some prompt engineering to correct this.

```
ran classify_meeting_intent: num outputs: 1
[bold]dynamic router returned node: summarize_sales_call[/bold]

Model response for summarize_sales_call: Summarize this sales call transcript, highlighting:
    - Key details
    - Client concerns
    - Action items
    - Next steps
    - Recommendations for improving the approach

    Transcript:Hello.
 Hi, May I speak with Vanessa Roberts?
 This is she.
 Good morning, Vanessa.
 This is Candice from Christchantalla.com.
 The reason for my call today is about the form you filled out on our website for our
 fiber plans.
 Have I caught you at a good time?
 For now, it's fine, but I might be busy soon.
 But we can talk about it now.
 Perfect.
 First, I'd like to quickly confirm your address and see if it's part of our serviceable locations.
 What we have here is 1-1-3-4-5 Keyes Bruce Way, San Antonio, Florida, 3-3-5-7-6.
 Is this correct?
 That's correct.
 Perfect.
 So this address is eligible for a fiber connection.
 We can provide you with a service.
 Would you like to know more about our available plans?
 Actually, I was just curious if my location is serviceable for fiber, but I'm not really actively looking to switch right now.
 I'm kind of happy with my current provider, so...
 I understand.
 This is just going to be a discovery call to see if there are areas that we might be able to help you out with.
 If I could just ask you real quick, how much is the average bid you're getting from your current provider?
 I'm not sure.
 I haven't really checked.
 And what's your current provider?
 BCNC.
 And what do you usually use your connection for?
 Browsing, gaming, streaming.
 All sorts of things, but mainly just browsing, face time, and streaming most of all.
 And are you totally happy with how the speed performs for each of these activities?
 Or do you think it could be better in some aspects?
 I would say totally happy, but mostly it's acceptable.
 When watching videos at 4K, there's definitely some lagging.
 It doesn't matter which device on our phones, on our TV.
 It was just freeze for a few seconds and then resume.
 I'm guessing it's a connection because it's definitely not this device.
 We just bought this 4K TV last month and it's been happening like clockwork every time we've watched 4K.
 Other than that, it's fine.
 You know, in streaming 4K videos, you'd normally need 25 megabits per second to effectively load without lagging.
 Anything lower than 25, it would definitely slow you down one way or another.
 Oh really, I wasn't exactly aware of that.
 Tell you what Vanessa, why don't we test your connection and see how much speed you're exactly getting?
 Because I think you're getting less than 25.
 Or if you are, chances are it's not a consistent 25 throughout the day.
 Are you in front of your computer right now?
 No, but now that you've raised that, I'm actually curious.
 Since I haven't really checked the speed before, huh?
 All right, let's just go to my computer. What do you want me to do?
 So if you could just open a browser and go to speedtask.net.
 Uh-huh, speedtask.net.
 Okay.
 And then click go once you're there.
 Okay.
 It says, oh wow, this thing's moving.
 Yep, that's the meter.
 The first number should be your download speed.
 Okay, it says 8.19 and there's just another wheel coming up.
 I guess that'll be the upload speed?
 Yep, that'll be the upload speed.
 Ah, so now it says 17.43.
 So 9.19 and 17.43.
 So definitely not enough, huh?
 No, definitely not enough to load 4K.
 You know when you're streaming videos,
 we need to look at your download speed.
 How much megabits you're downloading from the web per second?
 Okay.
 Your download speed is 8.19.
 That's good enough for 1080.
 But definitely not for 4K.
 Because you see, 4K technically is 4 times the size of 1080.
 So that's like 2 million pixels versus 8 million pixels per second.
 Oh, wow.
 So that explains the lagging then.
 Oh, definitely.
 There's no doubt about it.
 And that's what I'm calling you about today Vanessa.
 If you want to make the most out of your 4K TV,
 then have something much faster for you with an attractive pricing.
 Okay.
 You could get up to 800 megabits per second,
 which is roughly 100 times faster than your current download speed.
 Would you like to know more about the available plans?
 Well, that sounds great.
 But how much is it?
 It depends on the plan.
 But 800 megabits per second starts as low as $50 per month
 plus installation fees.
 Okay.
 I don't know.
 That's kind of too much for me.
 I don't think we're willing to pay that much.
 Our internet is a little slow,
 but it works for us most of the time.
 So I don't think so.
 Okay.
 How much do you pay monthly for your current connection?
 About $33 something like that.
 That's kind of like my price range.
 I don't think I want to be paying any higher than that.
 Okay.
 Okay.
 I getcha.
 Let me check here.
 Okay.
 Because I might have something more suitable for you.
 Oh.
 Okay.
 Okay.
 What if I tell you you can get 100 megabits per second
 for only $25 with free installation fee?
 $25 for how much speed?
 100 megabits.
 Not as fast as 800,
 but definitely 12 times faster than your current download speed.
 More than enough to play 4K and cheaper than $33.
 How does that sound?
 That's getting interesting.
 But first, I'm going to have to talk to my husband about it.
 I also have to go to work now.
 Can you call me later after work this evening?
 Sure.
 I can call you back at 6 or 7 this evening.
 Which do you prefer?
 7 should be fine.
 All right, then.
 Thank you for your time, Vanessa.
 I'll call you back at 7 in the evening today.
 Have a good day.
 You too.
 Bye.
 Hello.
 Hi, Vanessa.
 Good evening.
 This is Candice from Frischandtelecom.
 I'm following up with you as we agreed this morning
 to boost your internet speed
 for a 4K loading capability.
 Oh, yes, a little Candice.
 How are you?
 I'm doing fantastic.
 I'm just about to head home.
 I'm just doing my last calls for the day.
 How are you?
 Great, great.
 Thank you.
 I just finished work.
 And I have talked to my husband about it.
 We just have a few questions before we jump into it.
 Of course, sure.
 Go ahead.
 I'm a little unclear about the pricing.
 You said it was $25,
 which is a lot cheaper than my current provider
 and a lot faster too,
 which is what God is interested, you know?
 Sure.
 But what I'd like to know is,
 is $25 is going to be my regular monthly fee?
 Or is it only for a couple of months
 and I'll pay the regular price?
 Because if that is the case, we're not interested.
 Oh, $25 is actually already our regular pricing
 for 100 megabits per second fiber connection.
 Oh, and no additional fee on top of $25 after a few months?
 No additional fee because that's already the regular pricing
 and the installation is free as well.
 Oh, okay.
 Then great.
 Let's do it.
 Perfect.
 So to recap real quick,
 I've got you a plan for 100 megabits per second
 for a $25 monthly fee, no installation fee.
 Placing the order will only take a couple of minutes.
 I'll send you the link.
 Then I'll walk you through the steps to sign up.
 Are you ready?
 All right.
 I've just sent you the sign up link to your email
 Vanessa Roberts at gmail.com.
 Okay, let me check.
 Mm-hmm.
 I got it.
 Great.
 So just click on the link
 and fill out the required information
 and you should receive a confirmation email after that.
 Okay, I'm doing it now.
 Okay, I think I'm done.
 What's next?
 Now you can choose the plan and payment method
 and make sure to unclick the box
 next to the installation fee
 since you already have a landline installed.
 And click on submit order.
 Yes, click on submit order
 and that should complete the purchase.
 Okay, I think I'm done.
 Okay, let me check here.
 One moment.
 Congratulations.
 I see here that the order is successful.
 You will get your router within three business days
 and I will also send you a link to a video
 showing the step-by-step instructions
 on how to activate it.
 Thank you.
 But my service will continue working until
 we receive the router, right?
 Oh, definitely.
 Your service will continue working
 without any interruption.
 Perfect.
 We are honestly so excited about it
 and I hope we won't get disappointed.
 No, you won't be.
 When you're finally able to make the most out of your 4K TV
 with no more buffering,
 why would you be disappointed?
 Well, when you put it like that, it makes sense.
 And if you need help,
 you can always call a customer service hotline
 and they'll be there to answer
 your billing and technical questions.
 All right, Kenneth, thank you so much.
 And we'll just wait for the router.
 You have yourself a great evening.
 You too, Vanessa. Here is the summary of the sales call transcript:
**Key Details**

* The client, Vanessa Roberts, had filled out a form on the company's website (Christchantalla.com) inquiring about fiber plans for her address (1-1-3-4-5 Keyes Bruce Way, San Antonio, Florida, 3-3-5-7-6).
* The address is eligible for fiber connection, and the client's current provider is BCNC.
* Vanessa uses her connection for browsing, gaming, streaming, and experiences lagging when watching 4K videos.
* The client's current download speed is 8.19 Mbps, which is not sufficient for 4K streaming.

**Client Concerns**

* Vanessa is initially not looking to switch providers, but is interested in learning more about the company's plans.
* She is concerned about the cost of the plan, as she is currently paying around $33 per month and does not want to pay more.
* The client wants to know if the $25 per month plan for 100 Mbps is a promotional price or a regular monthly fee.

**Action Items**

* Candice, the sales representative, tests the client's current internet speed using speedtask.net.
* Candice explains the minimum speed requirements for 4K streaming (25 Mbps) and offers a plan with 100 Mbps for $25 per month.
* The client agrees to sign up for the plan after discussing it with her husband.

**Next Steps**

* Candice sends the client a link to sign up for the plan and walks her through the steps to complete the order.
* The client receives a confirmation email after completing the order.
* The company will send the client a router within three business days and provide instructions on how to activate it.
* The client's service will continue to work without interruption until the new router is activated.

**Recommendations for Improving the Approach**

* Candice effectively uses a discovery call approach to uncover the client's concerns and interests.
* However, she could have provided more information about the company's plans and pricing earlier in the conversation.
* Candice could have also offered to address the client's concerns about cost by offering a trial or a promotional period.
* The sales representative could have provided more details about the company's installation process and what to expect after signing up for the plan.
* It would be helpful to provide a clear and concise summary of the plan and pricing at the end of the conversation, rather than relying on the client to recall the details.

**Additional Observations**

* Candice, the sales representative, did a good job of:
        + Empathizing with the client's concerns and needs
        + Providing clear and concise information about Christchantalla's plans and pricing
        + Using data and facts to educate the client
        + Offering alternative plans and pricing options
* However, the sales representative could have:
        + Asked more questions about the client's current internet usage and needs
        + Provided more detailed information about the plans and pricing options
        + Clarified any confusion or misinformation early on in the call
        + Offered additional services or promotions to enhance the sale and increase customer satisfaction
```

## What

<!--
In a few sentences, please summarize the change to help reviewers.

Consider providing screenshots, logs, code or other visual aids to help the reviewer understand the approach taken.
-->
Added changes to the prompts in the compute functions:
1. **classify_meeting_intent**
2. **summarize_job_interview**
3. **summarize_sales_call**

## Testing

<!--
Please include steps used to verify the change.

Consider providing screenshots, logs, code or other visual aids to help the reviewer in their testing.
-->
Run **workflow.py** as given. Expected result:
![image](https://github.com/user-attachments/assets/3e7fee89-9770-494e-ae5c-7990f1dfadd4)


## Contribution Checklist

- [ ] If the python-sdk was changed, please run `make fmt` in `python-sdk/`.
- [ ] If the server was changed, please run `make fmt` in `server/`.
- [ ] Make sure all PR Checks are passing.
<!-- You can run the tests manually:
In `python-sdk/`, run the run `pip install -e .`, start the server and executor, `python test_graph_behaviours.py`.